+++ boot/pom.xml
@@ -51,18 +51,6 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-			<groupId>net.snaq</groupId>
-			<artifactId>dbpool</artifactId>
-			<version>7.0.1</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.jdbi</groupId>
-			<artifactId>jdbi3-core</artifactId>
-			<version>3.3.0</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
 			<groupId>mysql</groupId>
 			<artifactId>mysql-connector-java</artifactId>
 			<version>8.0.13</version>
