+++ boot/src/test/java/io/gridgo/boot/data/test/data/UserDAO.java
@@ -6,6 +6,7 @@ import org.joo.promise4j.Promise;
 
 import io.gridgo.boot.data.support.annotations.DataAccess;
 import io.gridgo.boot.data.support.annotations.JdbcProduce;
+import io.gridgo.boot.data.support.annotations.PojoMapper;
 import io.gridgo.boot.data.support.impl.JdbcDataAccessHandler;
 import io.gridgo.framework.support.Message;
 
@@ -21,6 +22,7 @@ public interface UserDAO {
     @JdbcProduce("insert into test_users (id, name) values (:1, :2)")
     public Promise<Message, Exception> add(int id, String name);
 
-    @JdbcProduce(value = "select * from test_users where id = :1", pojo = User.class)
+    @PojoMapper(User.class)
+    @JdbcProduce(value = "select * from test_users where id = :1")
     public Promise<List<User>, Exception> find(int id);
 }
